Update description of debug probe selector
