+++ changelog/changed-debug-probe-selector.md
@@ -1,0 +1,1 @@
+Moved `DebugprobeSelector` into its own module, remove some conversion traits to focus on `FromStr` implementation. Updated documentation to mention possibilty to select USB interface.
